Implementing the AdServe module, removing the ad boilerplate modal from the tracker page to a spearate component, exteding the communication service to support Ad tracking in the tracker page. Small bug fixes on the ContractService when disposing cache, removing Connect.js as we no longer need it for transactions. cleaning obsolete code from the index.js
