+++ src/Domain/Models/IncomingAdRequest.cs
@@ -1,0 +1,9 @@
+namespace Domain.Models
+{
+    public class IncomingAdRequest
+    {
+        public string Desktop { get; set; }
+        public string Tablet { get; set; }
+        public string Mobile { get; set; }
+    }
+}
