+++ src/Presentation/Components/TrackerComponents/AdvertisementComponent.razor
@@ -1,0 +1,30 @@
+
+@namespace SYNCWallet.Components.Advertisement
+
+<button class="btn btn-light-warning text-warning"  id="btnAdServe" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasBottom" aria-controls="offcanvasBottom" style="display: none;">Offcanvas Bottom</button>
+
+<div class="offcanvas offcanvas-bottom" tabindex="-1" id="offcanvasBottom" aria-labelledby="offcanvasBottomLabel">
+    <div class="offcanvas-body">
+        <div class="row">
+            <button type="button" class="btn-close text-reset" data-bs-dismiss="offcanvas" aria-label="Close"></button>
+        </div>
+        <div class="row">
+            @if (ServedAds != null)
+            {
+                @if (ServedAds.Count == 1)
+                {
+                    <img src="@ServedAds.FirstOrDefault().Desktop" width="100%" style="max-height: 349px;"/>
+                }
+                else
+                {
+                    @foreach (var ad in @ServedAds)
+                    {
+                        <div class="col-md-4">
+                            <img src="@ad.Desktop" width="100%" height="100%"/>
+                        </div>
+                    }
+                }
+            }
+        </div>
+    </div>
+</div>
