+++ src/Presentation/Components/TrackerComponents/PriceChangeComponent.razor
