+++ src/Presentation/Components/TrackerComponents/PriceChartComponent.razor
