+++ src/Presentation/wwwroot/js/Index.js
@@ -1,61 +1,4 @@
 var oneDelimiter = false;
-document.addEventListener('keyup', function (e) {
-    InputKey(e.key);
-});
-
-function removeFlex() {
-    document.getElementById("PageHeader").style.setProperty("display", "block");
-}
-
-function ReturnBack() {
-    window.history.go(-1);
-
-}
-
-
-function KeyboardToggle(state) {
-    if (state == "true")
-        document.getElementById("VirtualKeyboard").classList.add("visibleMenu");
-    else
-        document.getElementById("VirtualKeyboard").classList.remove("visibleMenu");
-
-    return "OK";
-}
-
-
-
-
-var selectedNetowrkId;
-var selectedChain = 97;
-var selectedContract = "0x78867BbEeF44f2326bF8DDd1941a4439382EF2A7";
-function MetworkChanged(args) {
-    var parsed = JSON.parse(args);
-    document.getElementById("networkLogo").setAttribute("src", parsed.Logo);
-    selectedChain = parsed.NetworkId;
-    selectedNetowrkId = undefined;
-    selectedContract = undefined;
-    console.log(args);
-
-}
-
-function CurrencyChanged(args) {
-    var parsed = JSON.parse(args);
-    document.getElementById("currencyLogo").setAttribute("src", parsed.Logo);
-    selectedNetowrkId = parsed.NetworkId;
-    selectedContract = parsed.ContractAddress;
-    console.log(args);
-    UpdateNetwork(selectedChain, selectedContract);
-}
-
-var cardAmount;
-function SetCardAmount() {
-    cardAmount = $("#price").html();
-    document.getElementById("submitTrigger").click();
-}
-
-function GetCardAmount() {
-    return cardAmount;
-}
 
 var isNotificationOpen = false;
 function OpenNotificationPanel() {
@@ -63,140 +6,6 @@ function OpenNotificationPanel() {
     document.getElementById("NotificationPanel").click();
 }
 
-
-
-document.addEventListener('touchstart', handleTouchStart, false);
-document.addEventListener('touchmove', handleTouchMove, false);
-
-var xDown = null;
-var yDown = null;
-
-function getTouches(evt) {
-    return evt.touches ||             // browser API
-        evt.originalEvent.touches; // jQuery
-}
-
-function handleTouchStart(evt) {
-    const firstTouch = getTouches(evt)[0];
-    xDown = firstTouch.clientX;
-    yDown = firstTouch.clientY;
-};
-
-function handleTouchMove(evt) {
-    if (!xDown || !yDown) {
-        return;
-    }
-
-    var xUp = evt.touches[0].clientX;
-    var yUp = evt.touches[0].clientY;
-
-    var xDiff = xDown - xUp;
-    var yDiff = yDown - yUp;
-
-    if (Math.abs(xDiff) > Math.abs(yDiff)) {/*most significant*/
-        if (xDiff > 0) {
-            /* right swipe */
-
-
-        } else {
-            /* left swipe */
-            if (isNotificationOpen) {
-                isNotificationOpen = false;
-                document.getElementById("closeHidden").click();
-
-            }
-        }
-    } else {
-        if (yDiff > 0) {
-            /* down swipe */
-        } else {
-            /* up swipe */
-        }
-    }
-    /* reset values */
-    xDown = null;
-    yDown = null;
-};
-
-
-function ToggleLock(args) {
-    if (args === 1) {
-        document.getElementById("unlockBtn").style.setProperty("display", "none");
-        document.getElementById("lockBtn").style.setProperty("display", "");
-        // Send .net command to unlock the door.
-    }
-    else {
-        document.getElementById("unlockBtn").style.setProperty("display", "");
-        document.getElementById("lockBtn").style.setProperty("display", "none");
-        // Send .net command to lock the door.
-
-    }
-    return "ok"
-}
-var resultContainer = document.getElementById('qr-reader-results');
-var lastResult, countResults = 0;
-
-function onScanSuccess(decodedText, decodedResult) {
-    if (decodedText !== lastResult) {
-        ++countResults;
-        lastResult = decodedText;
-        // Handle on success condition with the decoded message.
-        console.log(`Scan result ${decodedText}`, decodedResult);
-    }
-}
-
-function StartCam() {
-    //Call the API to retrive the door camera
-
-    var html5QrcodeScanner = new Html5QrcodeScanner(
-        "qr-reader", { fps: 10, qrbox: 450 });
-    html5QrcodeScanner.render(onScanSuccess);
-}
-
-
-function requestMedia() {
-
-    // Create request options
-    let options = {
-        audio: true,
-        video: true
-    };
-
-
-    // Warning: Below commented code causes "Illegal invocation error"
-    //// Set up proper function
-    //navigator.mediaDevices.getUserMedia =
-    //    navigator.getUserMedia ||
-    //    navigator.webkitGetUserMedia ||
-    //    navigator.mozGetUserMedia ||
-    //    navigator.msGetUserMedia;
-
-    try {
-        // Request user media
-        navigator.mediaDevices
-            .getUserMedia(options)
-            .then(gotLocalStream)
-            .catch(logError);
-    }
-    catch (e) {
-        logError(e);
-    }
-
-}
-
-// Receives local stream data
-function gotLocalStream(stream) {
-
-    // Get 'video' html control
-    var videoLocalPanel = document.getElementById("LOCAL_VIDEO_ID");
-
-    if ("srcObject" in videoLocalPanel) {
-        videoLocalPanel.srcObject = stream;
-    } else {
-        videoLocalPanel.src = window.URL.createObjectURL(stream);
-    }
-
-}
 
 function logError(e) {
 
@@ -445,4 +254,7 @@ function  TrackerCurrencyChanged(args)
     console.log(args);
 }
 
- 
+function OpenAdModule()
+{
+    document.getElementById("btnAdServe").click();
+}
